// portal-api/src/model/kanel/public/CapabilityPortal.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.CapabilityPortal */
export type CapabilityPortalId = string & { __brand: 'CapabilityPortalId' };

/** Represents the table public.CapabilityPortal */
export default interface CapabilityPortal {
  id: CapabilityPortalId;

  name: string | null;
}

/** Represents the initializer for the table public.CapabilityPortal */
export interface CapabilityPortalInitializer {
  id: CapabilityPortalId;

  name?: string | null;
}

/** Represents the mutator for the table public.CapabilityPortal */
export interface CapabilityPortalMutator {
  id?: CapabilityPortalId;

  name?: string | null;
}